+++ networktables.json
@@ -1,0 +1,18 @@
+[
+  {
+    "name": "/Preferences/ArmPosition",
+    "type": "double",
+    "value": 75.0,
+    "properties": {
+      "persistent": true
+    }
+  },
+  {
+    "name": "/Preferences/ArmP",
+    "type": "double",
+    "value": 50.0,
+    "properties": {
+      "persistent": true
+    }
+  }
+]
